services/emotion_service/requirements.txt
deepface==0.0.93
flask==2.3.2
pillow
# TensorFlow and dependencies for DeepFace
tensorflow-intel==2.16.1
ml-dtypes==0.3.1
tensorboard>=2.16,<2.17 